requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
app-model==0.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autopep8==2.3.1
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.2.post1
cachey==0.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dask==2024.10.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docstring_parser==0.16
docutils==0.21.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flexcache==0.3
flexparser==0.3.1
fonttools==4.54.1
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
HeapDict==1.0.1
hsluv==5.0.4
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.35.1
imagesize==1.4.1
in-n-out==0.2.1
ipycanvas==0.13.3
ipyevents==2.0.2
ipykernel==6.29.5
ipympl==0.9.4
ipython==8.28.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jupytext==1.16.4
kiwisolver==1.4.7
lazy_loader==0.4
locket==1.0.0
loguru==0.7.2
magicgui==0.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
napari==0.5.4
napari-assistant==0.4.10
napari-console==0.1.0
napari-plugin-engine==0.2.0
napari-segment-blobs-and-things-with-membranes==0.3.12
napari-simpleitk-image-processing==0.4.9
napari-skimage-regionprops==0.10.1
napari-svg==0.2.0
napari-time-slicer==0.5.0
napari-tools-menu==0.1.20
napari-workflows==0.2.11
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
npe2==0.7.7
numpy==1.26.4
numpydoc==1.8.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
Pint==0.24.3
platformdirs==4.3.6
pooch==1.8.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
psygnal==0.11.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyclesperanto_prototype==0.24.5
pycodestyle==2.12.1
pyconify==0.1.6
pycparser==2.22
pydantic==2.9.2
pydantic-compat==0.1.2
pydantic_core==2.23.4
Pygments==2.18.0
pyopencl==2024.3
PyOpenGL==3.1.7
pyparsing==3.1.4
pyperclip==1.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytools==2024.1.14
pytz==2024.2
pyvista==0.44.1
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
Send2Trash==1.8.3
setuptools==75.1.0
shellingham==1.5.4
SimpleITK==2.4.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
stackview==0.11.2
superqt==0.6.7
sympy==1.13.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
tomli_w==1.1.0
toolz==1.0.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trame==3.7.0
trame-client==3.4.0
trame-jupyter-extension==2.1.3
trame-server==3.2.3
transforms3d==0.4.2
triton==3.0.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
vispy==0.14.3
vtk==9.3.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.16.0
wslink==2.2.1
yarl==1.17.0
